magc: move magc dialogs from dlg_windows to magc_controls
